LPS-116144 Fix dataType property for numeric types

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/upgrade/v3_8_0/UpgradeDDMStructure.java
@@ -267,7 +267,7 @@ public class UpgradeDDMStructure extends UpgradeProcess {
 
 	private void _upgradeDecimalField(JSONObject jsonObject) {
 		jsonObject.put(
-			"dataType", "decimal"
+			"dataType", "double"
 		).put(
 			"type", "numeric"
 		).put(
@@ -466,7 +466,7 @@ public class UpgradeDDMStructure extends UpgradeProcess {
 
 	private void _upgradeNumberField(JSONObject jsonObject) {
 		jsonObject.put(
-			"dataType", "decimal"
+			"dataType", "double"
 		).put(
 			"type", "numeric"
 		).put(
